backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pyotp>=2.9.0
qrcode[pil]>=7.4.2
